+++ src/main/java/com/example/cinema/controller/comand/admin/DeleteSeanceCommand.java
@@ -1,7 +1,13 @@
 package com.example.cinema.controller.comand.admin;
 
+import com.example.cinema.controller.AccessStatus;
+import com.example.cinema.controller.ConfigurationManager;
 import com.example.cinema.controller.comand.ActionCommand;
-import com.example.cinema.model.service.MovieSeanceService;
+import com.example.cinema.controller.comand.common.GenerateMoviesPage;
+import com.example.cinema.controller.comand.common.GenerateSeancesPage;
+import com.example.cinema.model.entity.Role;
+import com.example.cinema.model.entity.User;
+import com.example.cinema.service.MovieSeanceService;
 
 import javax.servlet.http.HttpServletRequest;
 
@@ -13,11 +19,17 @@ public class DeleteSeanceCommand implements ActionCommand {
     @Override
     public String execute(HttpServletRequest req) {
         String page = null;
+        User user = (User)req.getSession().getAttribute("user");
+        if (!AccessStatus.isAdmin(user)) {
+            page = ConfigurationManager.getProperty("path.page.index");
+            return page;
+        }
         MovieSeanceService service = MovieSeanceService.getInstance();
 
         service.deleteSeance(Integer.parseInt(req.getParameter("seanceId")));
 
-        page = "/controller?command=getSeances";
+        GenerateSeancesPage command = new GenerateSeancesPage();
+        page = command.execute(req);
         return page;
     }
 }
